Move weather data type to ddmain package
